requirements.txt
Flask==2.0.2
sqlitedict==1.7.0
openai==0.27.0
gunicorn==20.1.0
python-dotenv==0.20.0
stripe==5.0.0
sendgrid==6.9.7
tiktoken==0.2.0
PyPDF2==3.0.1
pandas==1.5.2
Flask-Cors==3.0.10
beautifulsoup4==4.11.2
selenium==4.9.1
pyjwt==2.7.0
yt-dlp==2023.3.4
pycryptodome==3.15.0
